Move Source Academy game to features
