+++ public/assets/chapter0.1.txt
@@ -12,17 +12,17 @@ room
         show_dialogue(carpet) if checklist.visit
 
     objects
-        +door, /objects/yourCarpet/normal.png, 300, 300
+        + door, /objects/yourCarpet/normal.png, 300, 300
             show_dialogue(whoddaThought) if checklist.visit
             change_location(classroom) if checklist.visit 
-        +chair, /objects/yourChair/normal.png, 90, 90
-        +carpet, /objects/yourCarpet/normal.png, 90, 90
+        + chair, /objects/yourChair/normal.png, 90, 90
+        + carpet, /objects/yourCarpet/normal.png, 90, 90
 
     boundingBoxes
-        +bbox1, 10, 10, 30, 30
+        + bbox1, 10, 10, 30, 30
 
     characters
-        +scottie, Scottie, sad, left
+        + scottie, Scottie, sad, left
 
 hallway
     modes: talk, move
