+++ src/features/storySimulator/scenes/ObjectPlacement/ObjectPlacement.ts
@@ -4,8 +4,8 @@ import { AssetKey, AssetPath } from 'src/features/game/commons/CommonTypes';
 import GameInputManager from 'src/features/game/input/GameInputManager';
 import GameLayerManager from 'src/features/game/layer/GameLayerManager';
 import { Layer } from 'src/features/game/layer/GameLayerTypes';
+import SourceAcademyGame from 'src/features/game/SourceAcademyGame';
 import { mandatory } from 'src/features/game/utils/GameUtils';
-import SourceAcademyGame from 'src/pages/academy/game/subcomponents/SourceAcademyGame';
 
 import SSImageAssets from '../../assets/ImageAssets';
 import SSBackgroundManager from '../../background/SSBackgroundManager';
