+++ src/pages/academy/storySimulator/subcomponents/storySimulatorGame.ts
@@ -1,10 +1,9 @@
 import * as Phaser from 'phaser';
+import { screenSize } from 'src/features/game/commons/CommonConstants';
 import GameManager from 'src/features/game/scenes/gameManager/GameManager';
+import SourceAcademyGame, { GameType } from 'src/features/game/SourceAcademyGame';
 import MainMenu from 'src/features/storySimulator/scenes/mainMenu/MainMenu';
 import ObjectPlacement from 'src/features/storySimulator/scenes/ObjectPlacement/ObjectPlacement';
-
-import { screenSize } from '../../../../features/game/commons/CommonConstants';
-import SourceAcademyGame, { GameType } from '../../game/subcomponents/SourceAcademyGame';
 
 const config = {
   debug: true,
